refactor: import searchCategory from search component
